chore(riscv_transpiler): clear all clippy and rustc warnings (lib + tests)
332 diagnostics to zero. Real fixes dominate; 15 documented #[expect]s
(keccak seq!-unroll identity_op joining the existing eq_op/erasing_op;
Zimop MOP.R.n mask groupings verified against the spec encoding; 13
hot-loop needless_range_loop). 8 dead decoder helpers deleted with
cross-arch evidence; 12 jit-only MachineState offsets cfg-gated to their
x86_64+jit consumer. Test shape unchanged (34 passed / 4 pre-existing
fixture failures). jit/tests.rs (x86_64+jit-gated) untouched — flagged
for the CI-side clippy iteration.

+++ riscv_transpiler/src/replayer/instructions/mod.rs
@@ -27,7 +27,7 @@ pub(crate) fn touch_x0_with_ts<C: Counters, const TIMESTAMP_OFFSET: TimestampSca
     state: &mut State<C>,
 ) -> TimestampScalar {
     unsafe {
-        let reg = state.registers.get_unchecked_mut(0 as usize);
+        let reg = state.registers.get_unchecked_mut(0);
         let ts = reg.timestamp;
         debug_assert!(ts < (state.timestamp | TIMESTAMP_OFFSET));
         reg.timestamp = state.timestamp | TIMESTAMP_OFFSET;
